Implemented testing from terminal
